Refactor data initialization for modularity
Separate responsibilities for data initialization into dedicated
classes like CurrentAccountInitializer and SavingsAccountInitializer.
This adheres to the Single Responsibility Principle and enhances
maintainability and clarity in the data initialization process.
