+++ src/main/java/de/raywotrainings/banking/bankingservice/control/account/Entry.java
@@ -1,13 +1,15 @@
 package de.raywotrainings.banking.bankingservice.control.account;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 
 import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 
 @AllArgsConstructor
-@NoArgsConstructor
 @Getter
 @Setter
 @ToString
